requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Werkzeug==3.0.3
python-dotenv==1.0.1
Pillow==10.4.0
gunicorn==22.0.0
google-genai>=1.0.0
